requirements.txt
pandas==2.0.3
streamlit==1.39.0
streamlit_option_menu==0.4.0
openpyxl==3.1.5
pyodbc==5.2.0
toml==0.10.2
python-dotenv==1.0.1
